refactor package name from browser Connection to websocket follow team decision agreed on daily meeting 4-10-2020
